feat: Implementar sistema completo EducaGestor360 - Backend Spring Boot + Frontend React con autenticación JWT, gestión de usuarios, estudiantes, profesores y cursos. Sistema completamente funcional.

+++ backend/src/main/java/com/educagestor/entity/Course.java
@@ -272,6 +272,6 @@ public class Course extends BaseEntity {
         if (startDate == null || endDate == null) {
             return null;
         }
-        return startDate.until(endDate).getDays();
+        return (long) startDate.until(endDate).getDays();
     }
 }
